// front/src/models/user.model.ts
export interface User {
  id: number;
  name: string;
  username?: string;
  numberOfPosts?: number;
  email: string;
  address?: Address;
  phone?: string;
  website?: string;
  company?: Company;
}

interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: GeoLocation;
}

interface GeoLocation {
  lat: string;
  lng: string;
}

interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}